cleanup refactor and fixes

+++ ubuntu-14.04/scripts/cleanup.sh
@@ -9,10 +9,9 @@ rm /var/lib/dhcp/*
 
 echo "cleaning up udev rules"
 rm /etc/udev/rules.d/70-persistent-net.rules
-mkdir /etc/udev/rules.d/70-persistent-net.rules
+mkdir /etc/udev/rules.d/
 rm -rf /dev/.udev/
-rm /lib/udev/rules.d/75-persistent-net-generator.rules
-
+[ -f /lib/udev/rules.d/75-persistent-net-generator.rules ] && rm /lib/udev/rules.d/75-persistent-net-generator.rules
 echo "cleaning up minion_id for salt"
-rm /etc/salt/minion_id
+[ -f /etc/salt/minion_id ] &&  rm /etc/salt/minion_id
